requirements.txt
streamlit>=1.28.0
pandas>=1.5.0
plotly>=5.15.0
tavily-python>=0.3.0
openai>=1.3.0
sqlite3
beautifulsoup4>=4.12.0
requests>=2.31.0
python-dateutil>=2.8.0
sentence-transformers>=2.2.0
feedparser>=6.0.0
numpy>=1.24.0
scikit-learn>=1.3.0